in the end

+++ Project3Sorting/src/MergeSort.java
@@ -8,7 +8,7 @@
  */
 public class MergeSort {
 	
-	    void merge(byte arr[], int l, int m, int r)
+	    void merge(int arr[], int l, int m, int r)
 	    {
 	        int n1 = m - l + 1;
 	        int n2 = r - m;
